services/posts/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
python-dotenv
httpx  # <-- Añade esta línea